Reconstruction

+++ Assets/Frameworks/Service/Tridimensional.Puzzle.Service.Contract/LayoutContract.cs
@@ -9,7 +9,7 @@ namespace Tridimensional.Puzzle.Service.Contract
 	{
         public int Rows { get; set; }
         public int Columns { get; set; }
-        public float Width { get; set; }
-        public float Height { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
     }
 }
